+++ src/aeminium/runtime/task/implicit/ImplicitBlockingTask.java
@@ -5,7 +5,7 @@ import aeminium.runtime.task.RuntimeBlockingTask;
 
 public final class ImplicitBlockingTask<T extends ImplicitTask<T>> extends ImplicitTask<T> implements 	RuntimeBlockingTask {
 
-	public ImplicitBlockingTask(Body body, long hints) {
+	public ImplicitBlockingTask(Body body, short hints) {
 		super(body, hints);
 	}
 }
